DELIA-50319: wpeframework service crash
The in-process portion of the FireboltMediaPlayer cannot handle the
out-of-process portion stopped unexpectedly; either killed or crashing.
This is because it is assuming the error codes that are returned rather
than handing unexpected ones, such as ERROR_CONNECTION_CLOSED.

Reason for change: explicitly killing the FireboltMediaPlayer
out-of-process micro-service or having it crashes cause wpeframwork
to crash.
Risks: Low
Test Procedure: Launch FireboltMediaPlayer and use kill -9 on its
process.

Signed-off-by: Mikolaj Staworzynski <mikolaj.staworzynski@redembedded.com>

Original source of patch:
https://github.com/rdkcentral/rdkservices/pull/1336/commits/7e4bf2627dceb522818f63478b04e4dccfba8338

Removed one block of patch that we already have:

- Unregister(_T("setPosition"));
+ Unregister(_T("seekTo"));

Additional unregister for _mediaPlayerSink